requirements.txt
appdirs==1.4.4
async-generator==1.10
autopep8==1.5.7
awsebcli==3.20.2
botocore==1.21.36
cement==2.8.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.2
filelock==3.0.12
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
future==0.16.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
loguru==0.5.3
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.5.3
pathspec==0.5.9
pipenv==2021.5.29
promise==2.3
psycopg2-binary==2.9.2
pycodestyle==2.7.0
pydantic==1.8.2
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.18.0
python-multipart==0.0.5
# pywin32==301
PyYAML==5.4.1
requests==2.26.0
Rx==1.6.1
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==1.4.28
starlette==0.14.2
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.7
uvicorn==0.13.4
virtualenv==20.4.7
virtualenv-clone==0.5.7
watchgod==0.7
wcwidth==0.1.9
websockets==8.1
Werkzeug==2.0.2
win32-setctime==1.0.3
